Simplify equipment methods in Person
r6287
2021-10-13

## ISSUE
1. https://github.com/mars-sim/mars-sim/issues/463

## NEW
1. Create EquipmentOwner interface
   - Used by Person and Robot

## CHANGE
1. Modify various methods in Person.
2. Change how it works in getAmountResourceRemainingCapacity()
   and getTotalCapacity() in Person.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/task/CollectMinedMinerals.java
@@ -336,7 +336,7 @@ public class CollectMinedMinerals extends EVAOperation implements Serializable {
 				carryMass += suit.getAmountResourceRemainingCapacity(oxygenID);
 				carryMass += suit.getAmountResourceRemainingCapacity(waterID);
 			}
-			int carryCapacity = person.getCarryingCapacity();
+			double carryCapacity = person.getCarryingCapacity();
 			boolean canCarryEquipment = (carryCapacity >= carryMass);
 
 			result = (bagAvailable && canCarryEquipment);
