+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/task/DigLocal.java
@@ -269,7 +269,7 @@ implements Serializable {
         }
         
         if (!finishedCollecting) {
-        	int loadCap = person.getCarryingCapacity();
+        	double loadCap = person.getCarryingCapacity();
             if (totalCollected >= loadCap) {
             	totalCollected = loadCap;
     			finishedCollecting = true;
